Run locale-dependent tests with per-thread-locale
This is a fix to allow us to run Rust tests multithreaded. We use `wcrtomb` and
`mbrtowc` which are not thread-safe since they depend on the global locale.
Ideally we should instead use `wcrtomb_l` and `mbrtowc_l` and manage the locale
ourselves, but that is a larger code-change that's better suited for after the
rewrite is finished.

In the meantime we can employ per-thread-locale with `uselocale`, but that is
not available everywhere, in particular NetBSD does not support it. Therefore,
we conservatively only enable the per-thread locale on platforms we know support
it, and skip the tests that sets the locale on platforms that may not support
per-thread locale, to avoid interfering with other locale-sensitive tests.